Seperated Authentication Class into authentication.h and moved libray to Folders for both Client and Server
