+++ src/common/ChatAppClient/library.h
@@ -1,0 +1,17 @@
+#include <iostream>
+#include <thread>
+#include <cstring>
+#include <unistd.h>
+#include <sys/socket.h>
+#include <arpa/inet.h>
+#include <algorithm>
+#include <cctype>
+#include <netdb.h>
+#include <netinet/in.h>
+#include <sys/types.h>
+#include <fstream>
+#include <vector>
+#include <ifaddrs.h>
+#include <functional>
+
+using namespace std;
